requirements.txt
antlr4-python3-runtime==4.9.3
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
hydra-core==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
omegaconf==2.3.0
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
setuptools==69.0.3
six==1.16.0
